fix ioctl

+++ src/shared/platform/posix/nacl_host_desc.c
@@ -209,7 +209,7 @@ uintptr_t NaClHostDescMap(struct NaClHostDesc *d,
 
   if ((unsigned) mapbottom > (0xffffffffu - 256)) {
     errno = mapbottom;
-    mapbottom = (unsigned int) MAP_FAILED;
+    mapbottom = MAP_FAILED;
   } 
 
   /* MAP_FAILED is -1, so if we get that as our bottom 32 bits, we
